Fix exit issues and schema names with - in the name
On latest Ansible version the module was failing to do a proper
exit this commit fixes that.

It is now also possible to call the lib with optional scheme value,
previously the lib would split the key value (x.x.x-y-y-y)to
get the scheme so the scheme was x.x.x and the key was y-y-y
this doesn't hold when an application has - inside the scheme
name , to make it work with this kind of schemes,
now one can call the module with explicitly adding scheme and key.