Allow non-integer parts of GNOME version number
`gnome-shell --version` may return, e.g., `42.beta`, so we can't
assume that all parts of the version number are integers. Change how
the version number is parsed so that non-integer components are
treated as 0.

Another option would be to use `packaging.version.parse` to parse the
version numbers rather than parsing them ourselves, but that would
introduce to this module a dependency on `packaging.version`, which
would probably be gratuitous.