Work properly on recent Ubuntu releases
This commit fixes two problems with this module on recent Ubuntu releases:

1. Ubuntu GNOME sessions use `session=ubuntu` rather than `session=gnome`.
2. There are multiple `gnome-session-binary` processes so we only want to
   look at the first one.

The first issue is fixed by looking for `gnome-session-binary`
processes rather than `session=gnome` on the command line. Note,
however, that we only actually look for `gnome-session-b`, since the
`pgrep` command uses `/proc/\#/status`, which truncates the command
name at 15 characters.

The second issue is fixed by splitting the output of `pgrep` and only
using the first PID.